+++ src/AuthenticationContext.php
@@ -10,7 +10,18 @@ use Sirix\Mezzio\Authentication\Contract\TokenInterface;
 
 final readonly class AuthenticationContext implements AuthContextInterface
 {
-    public function __construct(private ?TokenInterface $token = null, private ?ActorInterface $actor = null) {}
+    private ?TokenInterface $token;
+    private ?ActorInterface $actor;
+
+    public function __construct(?TokenInterface $token = null, ?ActorInterface $actor = null)
+    {
+        $this->token = $token instanceof TokenInterface && $actor instanceof ActorInterface
+            ? $token
+            : null;
+        $this->actor = $actor instanceof ActorInterface && $token instanceof TokenInterface
+            ? $actor
+            : null;
+    }
 
     public function token(): ?TokenInterface
     {