+++ src/AuthenticationManager.php
@@ -10,6 +10,7 @@ use Sirix\Mezzio\Authentication\Contract\ActorInterface;
 use Sirix\Mezzio\Authentication\Contract\AuthContextInterface;
 use Sirix\Mezzio\Authentication\Contract\AuthManagerInterface;
 use Sirix\Mezzio\Authentication\Contract\TokenInterface;
+use Sirix\Mezzio\Authentication\Contract\TokenStorageInterface;
 use Sirix\Mezzio\Authentication\Contract\TokenStorageProviderInterface;
 use Sirix\Mezzio\Authentication\Contract\TokenTransportInterface;
 
@@ -17,29 +18,26 @@ final readonly class AuthenticationManager implements AuthManagerInterface
 {
     public function __construct(
         private TokenStorageProviderInterface $tokenStorageProvider,
-        private TokenTransportInterface $tokenTransport
+        private TokenTransportInterface $tokenTransport,
+        private ?string $storage = null,
     ) {}
 
-    public function login(array $payload, ?string $storage = null, ?int $expiresAt = null): TokenInterface
-    {
-        return null === $storage
-            ? $this->tokenStorageProvider
-                ->getDefaultStorage()
-                ->create($payload, $expiresAt)
-            : $this->tokenStorageProvider
-                ->getStorage($storage)
-                ->create($payload, $expiresAt)
-        ;
+    public function login(
+        ServerRequestInterface $serverRequest,
+        ResponseInterface $response,
+        array $payload,
+        ?int $expiresAt = null,
+    ): ResponseInterface {
+        $token = $this->storage()->create($payload, $expiresAt, $serverRequest);
+
+        return $this->tokenTransport->attach($response, $token);
     }
 
     public function logout(ServerRequestInterface $serverRequest, ResponseInterface $response): ResponseInterface
     {
         $token = $this->token($serverRequest);
         if ($token instanceof TokenInterface) {
-            $this->tokenStorageProvider
-                ->getStorage($token->getStorage())
-                ->delete($token, $serverRequest)
-            ;
+            $this->storage()->delete($token, $serverRequest);
         }
 
         return $this->tokenTransport->detach($response);
@@ -72,5 +70,12 @@ final readonly class AuthenticationManager implements AuthManagerInterface
         return $context instanceof AuthContextInterface
             ? $context
             : new AuthenticationContext();
+    }
+
+    private function storage(): TokenStorageInterface
+    {
+        return null === $this->storage
+            ? $this->tokenStorageProvider->getDefaultStorage()
+            : $this->tokenStorageProvider->getStorage($this->storage);
     }
 }